autenticação e autorização

+++ ApiDotNet6.Api/Controllers/PersonController.cs
@@ -1,5 +1,6 @@
 ﻿using ApiDotNet6.Application.DTOs;
 using ApiDotNet6.Application.Services.Interface;
+using ApiDotNet6.Domain.Authentication;
 using ApiDotNet6.Domain.FiltersDb;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -9,18 +10,28 @@ namespace ApiDotNet6.Api.Controllers
     [Authorize]
     [Route("api/[controller]")]
     [ApiController]
-    public class PersonController : ControllerBase
+    public class PersonController : BaseController
     {
-      private readonly IPersonService _personService;
-        public PersonController(IPersonService personService)   
+        private readonly IPersonService _personService;
+        private readonly ICurrentUser _currentUser;
+        private List<string> _permissionNeeded = new List<string>() { "admin" };
+        private readonly List<string> _permissionUser;
+
+        public PersonController(IPersonService personService, ICurrentUser currentUser)
         {
-            _personService  = personService;
+            _personService = personService;
+            _currentUser = currentUser;
+            _permissionUser = _currentUser?.Permissions?.Split(",")?.ToList() ?? new List<string>();
 
         }
 
         [HttpPost]
-        public async Task<ActionResult> PostAsync([FromBody] PersonDTO personDTO)
+        public async Task<IActionResult> PostAsync([FromBody] PersonDTO personDTO)
         {
+            _permissionNeeded.Add("CadastraPessoa");
+            if (!ValidPermission(_permissionUser, _permissionNeeded))
+                return ForBidden();
+
             var result = await _personService.CreateAsync(personDTO);
             if (result.IsSuccess)
                 return Ok(result);
@@ -29,10 +40,15 @@ namespace ApiDotNet6.Api.Controllers
         }
 
         [HttpGet]
-        public async Task<ActionResult> GetAsync()
+        public async Task<IActionResult> GetAsync()
         {
+            _permissionNeeded.Add("BuscaPessoa");
+            if (!ValidPermission(_permissionUser, _permissionNeeded))
+                return ForBidden();
+
+
             var result = await _personService.GetAsync();
-            if(result.IsSuccess)
+            if (result.IsSuccess)
                 return Ok(result);
 
             return BadRequest(result);
@@ -40,8 +56,12 @@ namespace ApiDotNet6.Api.Controllers
 
         [HttpGet]
         [Route("{id}")]
-        public async Task<ActionResult> GetByIdAsync(int id)
+        public async Task<IActionResult> GetByIdAsync(int id)
         {
+            _permissionNeeded.Add("BuscaPessoa");
+            if (!ValidPermission(_permissionUser, _permissionNeeded))
+                return ForBidden();
+
             var result = await _personService.GetByIdAsync(id);
             if (result.IsSuccess)
                 return Ok(result);
@@ -50,8 +70,12 @@ namespace ApiDotNet6.Api.Controllers
         }
 
         [HttpPut]
-        public async Task<ActionResult> UpdateAsync([FromBody] PersonDTO personDTO)
+        public async Task<IActionResult> UpdateAsync([FromBody] PersonDTO personDTO)
         {
+            _permissionNeeded.Add("EditaPessoa");
+            if (!ValidPermission(_permissionUser, _permissionNeeded))
+                return ForBidden();
+
             var result = await _personService.UpdateAsync(personDTO);
             if (result.IsSuccess)
                 return Ok(result);
@@ -61,8 +85,13 @@ namespace ApiDotNet6.Api.Controllers
 
         [HttpDelete]
         [Route("{id}")]
-        public async Task<ActionResult> DeleteAsync(int id)
+        public async Task<IActionResult> DeleteAsync(int id)
         {
+            _permissionNeeded.Add("DeletaPessoa");
+            if (!ValidPermission(_permissionUser, _permissionNeeded))
+                return ForBidden();
+
+
             var result = await _personService.DeleteAsync(id);
             if (result.IsSuccess)
                 return Ok(result);
@@ -75,7 +104,7 @@ namespace ApiDotNet6.Api.Controllers
         public async Task<ActionResult> GetPagedAsync([FromQuery] PersonFilterDb personFilterDb)
         {
             var result = await _personService.GetPagedAsync(personFilterDb);
-            if(result.IsSuccess)
+            if (result.IsSuccess)
                 return Ok(result);
 
             return BadRequest(result);
